# docs/pete_coach_openapi.yaml
openapi: 3.1.0
info:
  title: Pete-Eebot Coach API
  version: 1.0.0
  description: Coaching endpoints for a private Custom GPT action, including approximate nutrition logging.
servers:
  - url: https://api.myroadmapp.com
    description: HTTPS base URL that fronts Pete-Eebot.
security:
  - ApiKeyAuth: []
paths:
  /coach_state:
    get:
      operationId: get_coach_state
      summary: Get compact derived coaching state
      description: Default first call for the GPT. Returns readiness, trend deltas, load, data quality, goals, plan context, and missing subjective inputs. Browser sessions are assignment-scoped; the configured machine API key has trusted global profile-read scope.
      parameters:
        - name: date
          in: query
          required: true
          schema:
            type: string
            format: date
          description: Date in YYYY-MM-DD format. Use the most recent complete date for coaching decisions.
        - name: profile
          in: query
          required: false
          schema:
            type: string
          description: Optional profile slug. Browser sessions select an assigned default/first profile; the machine key selects the global default.
      responses:
        "200":
          description: Coach-ready state for the requested date.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CoachState"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/ProfileNotFound"
  /daily_summary:
    get:
      operationId: get_daily_summary
      summary: Get normalized daily metric summary
      description: Returns the daily summary with metric values, units, source labels, trust levels, and completeness metadata.
      parameters:
        - name: date
          in: query
          required: true
          schema:
            type: string
            format: date
          description: Date in YYYY-MM-DD format.
      responses:
        "200":
          description: Normalized daily summary.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DailySummary"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /recent_workouts:
    get:
      operationId: get_recent_workouts
      summary: Get recent running and strength workouts
      parameters:
        - name: days
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 90
            default: 14
        - name: end_date
          in: query
          required: false
          schema:
            type: string
            format: date
          description: Optional inclusive end date. Defaults to today on the server.
      responses:
        "200":
          description: Recent workouts split into running and strength arrays.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RecentWorkouts"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /nutrition/log-macros:
    post:
      operationId: log_nutrition_macros
      summary: Log approximate meal macros
      description: Stores ChatGPT-estimated meal macros in Pete-Eebot. Do not send images or raw wger entities.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NutritionLogRequest"
      responses:
        "200":
          description: Canonical stored nutrition event.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NutritionLogResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /nutrition/log-macros/{log_id}:
    patch:
      operationId: update_nutrition_log
      summary: Partially update a nutrition log entry
      parameters:
        - name: log_id
          in: path
          required: true
          schema:
            type: integer
            minimum: 1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NutritionLogPatchRequest"
      responses:
        "200":
          description: Updated nutrition event.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NutritionLogResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/BadRequest"
  /nutrition/daily-summary:
    get:
      operationId: get_nutrition_daily_summary
      summary: Get daily nutrition macro totals
      parameters:
        - name: date
          in: query
          required: true
          schema:
            type: string
            format: date
          description: Local date in YYYY-MM-DD format.
      responses:
        "200":
          description: Daily nutrition summary from approximate macro logs.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NutritionDailySummary"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /goal_state:
    get:
      operationId: get_goal_state
      summary: Get long-range goals and training maxes
      parameters:
        - name: profile
          in: query
          required: false
          schema:
            type: string
          description: Optional profile slug. Missing and existing-but-unassigned slugs use the same 404 response.
      responses:
        "200":
          description: Current goals and strength training max context.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GoalState"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/ProfileNotFound"
  /plan_context:
    get:
      operationId: get_plan_context
      summary: Get current plan phase and deload context
      parameters:
        - name: date
          in: query
          required: true
          schema:
            type: string
            format: date
          description: Date in YYYY-MM-DD format.
      responses:
        "200":
          description: Active plan context for the requested date.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlanContext"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /user_notes:
    get:
      operationId: get_user_notes
      summary: Get recent subjective notes
      description: Currently returns an empty not-configured payload until subjective notes persistence is added.
      parameters:
        - name: days
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 90
            default: 14
      responses:
        "200":
          description: Subjective notes payload.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserNotes"
        "401":
          $ref: "#/components/responses/Unauthorized"
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: Send PETEEEBOT_API_KEY in this header. Query-string API keys are not accepted. This trusted machine credential has profiles:read:any scope.
  responses:
    BadRequest:
      description: Invalid request parameter.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Unauthorized:
      description: Missing or invalid API key.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    ProfileNotFound:
      description: The profile is missing or is not assigned to the browser principal. These cases are intentionally indistinguishable.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message, correlation_id]
          properties:
            code:
              type: string
            message:
              type: string
            correlation_id:
              type: string
            details:
              type: object
              additionalProperties: true
    CoachState:
      type: object
      additionalProperties: true
      properties:
        date:
          type: string
          format: date
        profile:
          $ref: "#/components/schemas/UserProfile"
        summary:
          type: object
          properties:
            readiness_state:
              type: string
              enum: [green, amber, red]
            data_reliability_flag:
              type: string
              enum: [high, moderate, low]
            possible_underfueling_flag:
              type: boolean
            deload_due:
              type: boolean
        derived:
          type: object
          properties:
            weight_rate_pct_bw_per_week:
              type: number
              nullable: true
            sleep_debt_7d_minutes:
              type: number
              nullable: true
            rhr_delta_vs_28d_bpm:
              type: number
              nullable: true
            hrv_delta_vs_28d_ms:
              type: number
              nullable: true
            run_load_7d_km:
              type: number
              nullable: true
            strength_load_7d_kg:
              type: number
              nullable: true
        missing_subjective_inputs:
          type: array
          items:
            type: string
    DailySummary:
      type: object
      additionalProperties: true
      properties:
        date:
          type: string
          format: date
        metrics:
          type: object
          additionalProperties:
            $ref: "#/components/schemas/MetricValue"
        data_quality:
          type: object
          additionalProperties: true
          properties:
            status:
              type: string
            completeness_pct:
              type: number
            missing_fields:
              type: array
              items:
                type: string
    MetricValue:
      type: object
      properties:
        value:
          nullable: true
          oneOf:
            - type: number
            - type: string
            - type: boolean
        unit:
          type: string
          nullable: true
        source:
          type: string
        trust_level:
          type: string
          enum: [high, moderate, low]
        is_imputed:
          type: boolean
        data_quality:
          type: string
    RecentWorkouts:
      type: object
      additionalProperties: true
      properties:
        window:
          type: object
          additionalProperties: true
          properties:
            start_date:
              type: string
              format: date
            end_date:
              type: string
              format: date
            days:
              type: integer
        running:
          type: array
          items:
            type: object
            additionalProperties: true
            properties:
              workout_id:
                type: string
              workout_date:
                type: string
                format: date
              workout_type:
                type: string
              duration_sec:
                type: number
              total_distance_km:
                type: number
              pace_min_per_km:
                type: number
              avg_hr:
                type: number
              max_hr:
                type: number
              elevation_gain_m:
                type: number
        strength:
          type: array
          items:
            type: object
            additionalProperties: true
            properties:
              workout_date:
                type: string
                format: date
              exercise_id:
                type: integer
              exercise_name:
                type: string
              sets:
                type: integer
              reps:
                type: integer
              volume_kg:
                type: number
              top_weight_kg:
                type: number
              avg_rir:
                type: number
        data_quality:
          type: object
          additionalProperties: true
          properties:
            running_available:
              type: boolean
            strength_available:
              type: boolean
    NutritionLogRequest:
      type: object
      required: [protein_g, carbs_g, fat_g]
      properties:
        protein_g:
          type: number
          minimum: 0
        carbs_g:
          type: number
          minimum: 0
        fat_g:
          type: number
          minimum: 0
        timestamp:
          type: string
          format: date-time
          description: Approximate meal time. If omitted, Pete-Eebot uses server-local current time.
        source:
          type: string
          default: photo_estimate
        context:
          type: string
          nullable: true
          description: Optional coaching context such as post_run or evening_meal.
        confidence:
          type: string
          enum: [low, medium, high]
          default: medium
        meal_label:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
        client_event_id:
          type: string
          nullable: true
          description: Optional idempotency key generated by the caller.
    NutritionLogPatchRequest:
      type: object
      additionalProperties: false
      properties:
        protein_g: { type: number, minimum: 0 }
        carbs_g: { type: number, minimum: 0 }
        fat_g: { type: number, minimum: 0 }
        alcohol_g: { type: number, minimum: 0, nullable: true }
        fiber_g: { type: number, minimum: 0, nullable: true }
        estimated_total_calories: { type: number, minimum: 0, nullable: true }
        source: { type: string, nullable: true }
        context: { type: string, nullable: true }
        confidence: { type: string, enum: [low, medium, high] }
        meal_label: { type: string, nullable: true }
        notes: { type: string, nullable: true }
        timestamp: { type: string, format: date-time }
        eaten_at: { type: string, format: date-time }
    NutritionLogResponse:
      type: object
      additionalProperties: true
      properties:
        id:
          type: integer
        eaten_at:
          type: string
          format: date-time
        local_date:
          type: string
          format: date
        protein_g:
          type: number
        carbs_g:
          type: number
        fat_g:
          type: number
        calories_est:
          type: number
        source:
          type: string
        context:
          type: string
          nullable: true
        confidence:
          type: string
          enum: [low, medium, high]
        duplicate:
          type: boolean
        warnings:
          type: array
          items:
            type: string
    NutritionDailySummary:
      type: object
      additionalProperties: true
      properties:
        date:
          type: string
          format: date
        protein_g:
          type: number
        carbs_g:
          type: number
        fat_g:
          type: number
        calories_est:
          type: number
        meals_logged:
          type: integer
        source_breakdown:
          type: object
          additionalProperties:
            type: integer
        confidence_breakdown:
          type: object
          additionalProperties:
            type: integer
        data_quality:
          type: object
          additionalProperties: true
    GoalState:
      type: object
      additionalProperties: true
      properties:
        profile:
          $ref: "#/components/schemas/UserProfile"
        running_goal:
          type: object
          additionalProperties: true
          properties:
            target_race:
              type: string
            race_date:
              type: string
              format: date
            target_time:
              type: string
              nullable: true
            weight_loss_target_kg:
              type: number
              nullable: true
        body_composition_goal:
          type: object
          additionalProperties: true
          properties:
            goal_weight_kg:
              type: number
              nullable: true
            height_cm:
              type: integer
              nullable: true
        strength:
          type: object
          additionalProperties: true
          properties:
            training_maxes_kg:
              type: object
              additionalProperties:
                type: number
            training_max_last_measured_at:
              type: string
              format: date
              nullable: true
        performance_anchors:
          type: object
          additionalProperties: true
          properties:
            current_half_marathon_marker:
              type: string
            historical_half_marathon_pb:
              type: string
            sub_3_marathon_half_equivalent:
              type: string
    UserProfile:
      type: object
      additionalProperties: false
      properties:
        id:
          type: integer
          nullable: true
        slug:
          type: string
        display_name:
          type: string
        date_of_birth:
          type: string
          format: date
          nullable: true
        height_cm:
          type: integer
          nullable: true
        goal_weight_kg:
          type: number
          nullable: true
        timezone:
          type: string
        is_default:
          type: boolean
        is_active:
          type: boolean
    PlanContext:
      type: object
      additionalProperties: true
      properties:
        date:
          type: string
          format: date
        current_week_number:
          type: integer
          nullable: true
        mesocycle:
          type: string
        strength_phase:
          type: string
        deload_due:
          type: boolean
        data_quality:
          type: string
    UserNotes:
      type: object
      additionalProperties: true
      properties:
        window_days:
          type: integer
        notes:
          type: array
          items:
            type: object
            additionalProperties: true
            properties:
              date:
                type: string
                format: date
              note_type:
                type: string
              text:
                type: string
        data_quality:
          type: object
          additionalProperties: true
          properties:
            status:
              type: string
            message:
              type: string